Add Sale CRUD feature

+++ src/main/java/br/edu/ufcg/cccpharma/sale/SaleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface SaleRepository  extends JpaRepository<Sale, Long>{
 	
-	Sale findById(int Id);
+	Sale findById(long Id);
 
 }
